+++ src/bundle-browserify.js
@@ -1,0 +1,2 @@
+var mix = require('./modules/mix');
+module.exports = mix;